Combine Game script to GameManager
