+++ Assets/Scenes/MainScene.unity
@@ -15569,9 +15569,17 @@ MonoBehaviour:
   m_GameObject: {fileID: 963194225}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: cd81b87af86f2c449b5a97cb48a42262, type: 3}
+  m_Script: {fileID: 11500000, guid: c8a0f6bf6bd3ebd4da852e40984befb0, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  score: {fileID: 945856401}
+  ranking: []
+  timeText: {fileID: 456877033}
+  playerID: {fileID: 2859192194950675887, guid: 50148b92a3dd60940993ec8932b761ec,
+    type: 3}
+  idBox: {fileID: 135178802}
+  gameoverBackground: {fileID: 0}
+  gameResult: {fileID: 0}
 --- !u!1001 &963713320
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -31822,57 +31830,6 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.23322739, y: 0.13720722, z: 0.04100569}
   m_Center: {x: 0, y: 0.06860361, z: 0}
---- !u!1 &1848615840
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1848615842}
-  - component: {fileID: 1848615841}
-  m_Layer: 0
-  m_Name: GameManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &1848615841
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1848615840}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c8a0f6bf6bd3ebd4da852e40984befb0, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  score: {fileID: 945856401}
-  ranking: []
-  timeText: {fileID: 456877033}
-  playerID: {fileID: 2859192194950675887, guid: 50148b92a3dd60940993ec8932b761ec,
-    type: 3}
-  idBox: {fileID: 135178802}
-  gameoverBackground: {fileID: 0}
-  gameResult: {fileID: 0}
---- !u!4 &1848615842
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1848615840}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 96.850174, y: 310.0613, z: -33.680126}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 8
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1852050582
 PrefabInstance:
   m_ObjectHideFlags: 0
