--- Assets/Scripts/Game.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cd81b87af86f2c449b5a97cb48a42262
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
